fix(js): 调整 API 模型位置 + 小优化t
